* libamf/buffer.{h,cpp}: Cleanup minor bugs found by newly 	expanded test case. Rename empty() to clear(). 	* libamf/amf.cpp: Don't take 1 off the end of the string. Comment 	out some stuff that will be deleted later, it's been moved to the 	Element class. 	* libamf/element.cpp: Add makeNumber that takes a variable name. 	* testsuite/libamf.all/test_buffer.cpp: Add many new tests for all 	methods. This makes it much easier to more thoroughly test for 	memory problems. 	* testsuite/libamf.all/test_lc.cpp: Clean up error handling. 	* libbase/shm.cpp: Handle permission errors better. 	* libbase/gmemory.h, memory.cpp: Use size_t instead of int.

+++ libbase/gmemory.h
@@ -48,7 +48,7 @@ public:
         int fordblks; // total free space
     };
     Memory();
-    Memory(int size);
+    Memory(size_t size);
     ~Memory();
 
     // Start collecting statistics. This can effect performance
